apps/api/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart
google-generativeai>=0.4.1
pydantic>=2.7.0
python-dotenv>=1.0.0
